Update: Completed rough outline of 'Models', Added prototype of 'resolvers.js' & 'typeDefs.js'
