+++ server/models/01-AboutMe.js
@@ -5,13 +5,21 @@ const aboutMeSchema = new Schema({
     firstName: {
         type: String,
         required: true,
+        maxlength: 20,
         trim: true,
     },
     lastName: {
         type: String,
         required: true,
+        maxlength: 20,
         trim: true,
     },
+    introduction: {
+        type: String,
+        required: false,
+        trim: true,
+    },
+    
 });
