--- server/models/04-Projects.js
@@ -1,17 +1,0 @@
-const { Schema, model } = require("mongoose");
-
-
-const projectSchema = new Schema({
-    username: {
-
-    },
-    email: {
-
-    },
-});
-
-
-
-const Projects = model("Project", projectSchema);
-
-module.exports = Projects;
