--- server/models/05-ContactMe.js
@@ -1,17 +1,0 @@
-const { Schema, model } = require("mongoose");
-
-
-const contactMeSchema = new Schema({
-    username: {
-
-    },
-    email: {
-
-    },
-});
-
-
-
-const ContactMe = model("ContactMe", contactMeSchema);
-
-module.exports = ContactMe;
