+++ server/models/index.js
@@ -3,11 +3,11 @@ const User = require('./User');
 const AboutMe = require('./01-AboutMe');
 const Education = require('./02-Education');
 const Skills = require('./03-Skills');
-const Projects = require('./04-Projects');
-const ContactMe = require('./05-ContactMe');
+const Experience = require('./04-Experience');
+const Projects = require('./05-Projects');
+const ContactMe = require('./06-ContactMe');
 
 
 
 
-
-module.exports = { User, AboutMe, Education, Skills, Projects, ContactMe };
+module.exports = { User, AboutMe, Education, Skills, Experience, Projects, ContactMe };
